fix(mui-file-selector): remove form and make input standalone
