+++ packages/file-selector/src/lib/FileSelector.stories.tsx
@@ -1,9 +1,13 @@
 // Each exported component in the package should have its own stories file
 import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { FormProvider, useForm } from 'react-hook-form';
+import { QueryClient, QueryClientProvider, useQueryClient } from '@tanstack/react-query';
+import type { default as Upload } from '@availity/upload-core';
 import { Paper } from '@availity/mui-paper';
 import { Alert, AlertProps, AlertTitle } from '@availity/mui-alert';
+import { Grid } from '@availity/mui-layout';
+import { Button } from '@availity/mui-button';
 import Collapse from '@mui/material/Collapse';
 
 import { FileSelector, FileSelectorProps } from './FileSelector';
@@ -49,15 +53,46 @@ const DismissableAlert = (props: AlertProps) => {
 
 export const _FileSelector: StoryObj<typeof FileSelector> = {
   render: (props: FileSelectorProps) => {
+    const methods = useForm({
+      defaultValues: {
+        [props.name]: [] as File[],
+      },
+    });
+
+    const client = useQueryClient();
+
+    const files = methods.watch(props.name);
+
+    const handleOnSubmit = (values: Record<string, File[]>) => {
+      if (values[props.name].length === 0) return;
+
+      const queries = client.getQueriesData<Upload>(['upload']);
+      const uploads = [];
+      for (const [, data] of queries) {
+        if (data) uploads.push(data);
+      }
+    };
+
     return (
       <Paper sx={{ padding: '2rem' }}>
-        <FileSelector {...props}>
-          <DismissableAlert severity="warning">
-            <AlertTitle>Make an Appeal</AlertTitle>
-            This is an example alert. It is not part of the component. `children` you pass to the component will show up
-            here.
-          </DismissableAlert>
-        </FileSelector>
+        <FormProvider {...methods}>
+          <form onSubmit={methods.handleSubmit(handleOnSubmit)}>
+            <FileSelector {...props}>
+              <DismissableAlert severity="warning">
+                <AlertTitle>Make an Appeal</AlertTitle>
+                This is an example alert. It is not part of the component. `children` you pass to the component will
+                show up here.
+              </DismissableAlert>
+            </FileSelector>
+            {files.length > 0 && (
+              <Grid xs={12} justifyContent="end" display="flex" paddingTop={2.5}>
+                <Button type="submit" sx={{ marginLeft: 'auto', marginRight: 0 }}>
+                  Submit
+                </Button>
+              </Grid>
+            )}
+          </form>
+        </FormProvider>
       </Paper>
     );
   },
